requirements.txt
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
streamlit>=1.28.0
asyncio
typing-extensions>=4.0.0
pydantic>=2.0.0
